requirements.txt
discord.py
openai
requests
groq
python-dotenv
flask
